requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.0
pillow==12.0.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
